+++ peach_inv/coins.h
@@ -2,12 +2,12 @@
 using namespace std;
 class Coins{
 private:
-    int value;
+    float value;
 public:
     Coins();
-    Coins(int x);
-    int get_value();
-    void set_value(int val);
+    Coins(float x);
+    float get_value();
+    void set_value(float val);
     Coins operator+(Coins o);
 
 };
